refactor: move queries to TS
