+++ src/get-queries-for-element.ts
@@ -10,14 +10,17 @@ import * as defaultQueries from './queries'
  * @param {Object} initialValue for reducer
  * @returns {FuncMap} returns object of functions bound to container
  */
+
+type BoundFunction = {[key: string]: Function}
+
 function getQueriesForElement(
-  element,
-  queries = defaultQueries,
-  initialValue = {},
+  element: HTMLElement,
+  queries: BoundFunction = defaultQueries,
+  initialValue: BoundFunction = {},
 ) {
   return Object.keys(queries).reduce((helpers, key) => {
     const fn = queries[key]
-    helpers[key] = fn.bind(null, element)
+    helpers[key] = fn.bind(null, element) as Function
     return helpers
   }, initialValue)
 }
