+++ src/get-user-code-frame.ts
@@ -1,21 +1,25 @@
 // We try to load node dependencies
-let chalk = null
-let readFileSync = null
-let codeFrameColumns = null
+let chalk: {dim: Function} | null = null
+let readFileSync: Function | null = null
+let codeFrameColumns: Function | null = null
 
 try {
-  const nodeRequire = module && module.require
+  // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+  const nodeRequire = module?.require
 
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment,@typescript-eslint/no-unsafe-member-access
   readFileSync = nodeRequire.call(module, 'fs').readFileSync
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access,@typescript-eslint/no-unsafe-assignment
   codeFrameColumns = nodeRequire.call(module, '@babel/code-frame')
     .codeFrameColumns
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
   chalk = nodeRequire.call(module, 'chalk')
 } catch {
   // We're in a browser environment
 }
 
 // frame has the form "at myMethod (location/to/my/file.js:10:2)"
-function getCodeFrame(frame) {
+function getCodeFrame(frame: string) {
   const locationStart = frame.indexOf('(') + 1
   const locationEnd = frame.indexOf(')')
   const frameLocation = frame.slice(locationStart, locationEnd)
@@ -29,12 +33,14 @@ function getCodeFrame(frame) {
 
   let rawFileContents = ''
   try {
-    rawFileContents = readFileSync(filename, 'utf-8')
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+    rawFileContents = readFileSync?.(filename, 'utf-8')
   } catch {
     return ''
   }
 
-  const codeFrame = codeFrameColumns(
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+  const codeFrame = codeFrameColumns?.(
     rawFileContents,
     {
       start: {line, column},
@@ -44,7 +50,7 @@ function getCodeFrame(frame) {
       linesBelow: 0,
     },
   )
-  return `${chalk.dim(frameLocation)}\n${codeFrame}\n`
+  return `${chalk?.dim(frameLocation)}\n${codeFrame}\n`
 }
 
 function getUserCodeFrame() {
@@ -53,13 +59,15 @@ function getUserCodeFrame() {
   if (!readFileSync || !codeFrameColumns) {
     return ''
   }
-  const err = new Error()
-  const firstClientCodeFrame = err.stack
+  const err: Error = new Error()
+  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+  // @ts-expect-error
+  const firstClientCodeFrame: string | undefined = err.stack
     .split('\n')
     .slice(1) // Remove first line which has the form "Error: TypeError"
     .find(frame => !frame.includes('node_modules/')) // Ignore frames from 3rd party libraries
 
-  return getCodeFrame(firstClientCodeFrame)
+  return getCodeFrame(firstClientCodeFrame ?? '')
 }
 
 export {getUserCodeFrame}
