+++ types/matches.d.ts
@@ -10,7 +10,7 @@ export type Matcher = MatcherFunction | RegExp | number | string
 
 // Get autocomplete for ARIARole union types, while still supporting another string
 // Ref: https://github.com/microsoft/TypeScript/issues/29729#issuecomment-505826972
-export type ByRoleMatcher = ARIARole | MatcherFunction | {}
+export type ByRoleMatcher = ARIARole | Matcher | MatcherFunction
 
 export type NormalizerFn = (text: string) => string
